Raycast

Documentation for Raycast.

Provides a simple and fast method to run the Bresenham algorithm for raycasting in 3D.

Usage

using Raycast

p0 = (0,0,0)      # any tuple of integers will work
p1 = (10, 5, -3)  # any tuple of integers will work

prob = BresenhamProblem(p0, p1)
for p in prob
  @show p
end

p = (0, 0, 0)
p = (1, 1, 0)
p = (2, 1, -1)
p = (3, 2, -1)
p = (4, 2, -1)
p = (5, 3, -2)
p = (6, 3, -2)
p = (7, 4, -2)
p = (8, 4, -2)
p = (9, 5, -3)
p = (10, 5, -3)

When you construct BresenhamProblem it returns an iterator, so you can do things like

collect(prob)

11-element Vector{Tuple{Int64, Int64, Int64}}:
 (0, 0, 0)
 (1, 1, 0)
 (2, 1, -1)
 (3, 2, -1)
 (4, 2, -1)
 (5, 3, -2)
 (6, 3, -2)
 (7, 4, -2)
 (8, 4, -2)
 (9, 5, -3)
 (10, 5, -3)

or

length(prob)

11

The benefit of using the iterator interface is that it is non-allocating, and can support early termination. For instance,

using LinearAlgebra
p0 = (0,0,0)
p1 = (100000000, 5, -3)

prob = BresenhamProblem(p0, p1)
begin
  for p in prob
     @show p
     if (norm(p) > 10)
       println("far!")
       break
     end
  end
end

p = (0, 0, 0)
p = (1, 0, 0)
p = (2, 0, 0)
p = (3, 0, 0)
p = (4, 0, 0)
p = (5, 0, 0)
p = (6, 0, 0)
p = (7, 0, 0)
p = (8, 0, 0)
p = (9, 0, 0)
p = (10, 0, 0)
p = (11, 0, 0)
far!

Alternatively, if you don't like the iterator interface, you can directly call

p1 = (10, 5, -3)  # any tuple of integers will work
bresenham(p0, p1)

11-element Vector{Tuple{Int64, Int64, Int64}}:
 (0, 0, 0)
 (1, 1, 0)
 (2, 1, -1)
 (3, 2, -1)
 (4, 2, -1)
 (5, 3, -2)
 (6, 3, -2)
 (7, 4, -2)
 (8, 4, -2)
 (9, 5, -3)
 (10, 5, -3)

to get the entire vector. Note, this is slower than collect(prob) (and gives the same result) so there is really no reason to use this.

WARNING

The code is type-stable. Since sometimes we double some quantities, this can cause a integer overflow.

Therefore, it is safest to ensure your input is of type Int64. There is no meaningful loss of speed if you do this. An error will be thrown if it is likely for your types to overflow

Credits

The code (and logic) is taken from https://www.geeksforgeeks.org/bresenhams-algorithm-for-3-d-line-drawing/

Index

Raycast.BresenhamProblem

Raycast.Point2

Raycast.Point3

Raycast.bresenham

Raycast.bresenham

API

Raycast.BresenhamProblem — Type

BresenhamProblem(start_point::Point3, end_point::Point3)

returns an iterator that will cast a ray from startpoint to endpoint.

Examples

julia> p0 = (0,0,0)
julia> p1 = (10,0,0)

julia> for p in BresenhamProblem(p0, p1)
         @show p
       end

Raycast.Point2 — Type

Point2{T} is an alias for Tuple{T,T}

Raycast.Point3 — Type

Point3{T} is an alias for Tuple{T,T,T}

Raycast.bresenham — Method

bresenham(start_point::Point3, end_point::Point3)

returns a Vector{Point3} of the pixels touched by the ray connecting start point to end point

Raycast.bresenham — Method

bresenham(start_point::Point2, end_point::Point2)

returns a Vector{Point2} of the pixels touched by the ray connecting start point to end point